backend/requirements.txt
flask
chromadb
langchain
langchain-google-genai
pypdf
python-dotenv
google-generativeai